Update: ICU manage

+++ usb/usb_io.hpp
@@ -101,7 +101,7 @@ namespace device {
 		void enable_intr(bool ena = true) noexcept
 		{
 ///			format("USB0: enable_intr(%d)...\n") % static_cast<int>(ena);
-			ICU::IER.enable(static_cast<uint8_t>(ivec_), ena);
+			ICU::IER.enable(ivec_, ena);
 		}
 	};
